# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Noor Islamic App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserPreferencesLanguage =
  (typeof UserPreferencesLanguage)[keyof typeof UserPreferencesLanguage];

export const UserPreferencesLanguage = {
  en: "en",
  ar: "ar",
  fr: "fr",
  de: "de",
} as const;

export interface UserPreferences {
  language: UserPreferencesLanguage;
  calculationMethod?: string;
  theme?: string;
}

export type UpdatePreferencesRequestLanguage =
  (typeof UpdatePreferencesRequestLanguage)[keyof typeof UpdatePreferencesRequestLanguage];

export const UpdatePreferencesRequestLanguage = {
  en: "en",
  ar: "ar",
  fr: "fr",
  de: "de",
} as const;

export interface UpdatePreferencesRequest {
  language?: UpdatePreferencesRequestLanguage;
  calculationMethod?: string;
  theme?: string;
}

export interface AdhkarProgress {
  id: number;
  userId: number;
  dhikrId: string;
  count: number;
  completed: boolean;
  date: string;
}

export interface UpdateAdhkarProgressRequest {
  count: number;
  completed: boolean;
}

export type FavoriteType = (typeof FavoriteType)[keyof typeof FavoriteType];

export const FavoriteType = {
  surah: "surah",
  dhikr: "dhikr",
  verse: "verse",
} as const;

export interface Favorite {
  id: number;
  userId: number;
  type: FavoriteType;
  referenceId: string;
  title: string;
  createdAt: string;
}

export type AddFavoriteRequestType =
  (typeof AddFavoriteRequestType)[keyof typeof AddFavoriteRequestType];

export const AddFavoriteRequestType = {
  surah: "surah",
  dhikr: "dhikr",
  verse: "verse",
} as const;

export interface AddFavoriteRequest {
  type: AddFavoriteRequestType;
  referenceId: string;
  title: string;
}

export type CalendarEventType =
  (typeof CalendarEventType)[keyof typeof CalendarEventType];

export const CalendarEventType = {
  personal: "personal",
  islamic: "islamic",
  reminder: "reminder",
} as const;

export interface CalendarEvent {
  id: number;
  userId: number;
  title: string;
  description?: string;
  date: string;
  hijriDate?: string;
  type: CalendarEventType;
  createdAt: string;
}

export type CreateEventRequestType =
  (typeof CreateEventRequestType)[keyof typeof CreateEventRequestType];

export const CreateEventRequestType = {
  personal: "personal",
  islamic: "islamic",
  reminder: "reminder",
} as const;

export interface CreateEventRequest {
  title: string;
  description?: string;
  date: string;
  hijriDate?: string;
  type: CreateEventRequestType;
}