# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Noor Islamic App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: user
    description: User profile and preferences
  - name: adhkar
    description: Adhkar progress tracking
  - name: favorites
    description: User favorites
  - name: events
    description: Islamic calendar events
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /user/preferences:
    get:
      operationId: getPreferences
      tags: [user]
      summary: Get user preferences
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    put:
      operationId: updatePreferences
      tags: [user]
      summary: Update user preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesRequest"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"

  /adhkar/progress:
    get:
      operationId: getAdhkarProgress
      tags: [adhkar]
      summary: Get adhkar progress
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Adhkar progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdhkarProgress"

  /adhkar/progress/{dhikrId}:
    put:
      operationId: updateAdhkarProgress
      tags: [adhkar]
      summary: Update dhikr progress
      security:
        - bearerAuth: []
      parameters:
        - name: dhikrId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdhkarProgressRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdhkarProgress"

  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get user favorites
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Favorites list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Favorite"
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a favorite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteRequest"
      responses:
        "201":
          description: Favorite added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"

  /favorites/{id}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a favorite
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get user calendar events
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"

  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update a calendar event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete a calendar event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, createdAt]

    UserPreferences:
      type: object
      properties:
        language:
          type: string
          enum: [en, ar, fr, de]
        calculationMethod:
          type: string
        theme:
          type: string
      required: [language]

    UpdatePreferencesRequest:
      type: object
      properties:
        language:
          type: string
          enum: [en, ar, fr, de]
        calculationMethod:
          type: string
        theme:
          type: string

    AdhkarProgress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dhikrId:
          type: string
        count:
          type: integer
        completed:
          type: boolean
        date:
          type: string
          format: date
      required: [id, userId, dhikrId, count, completed, date]

    UpdateAdhkarProgressRequest:
      type: object
      properties:
        count:
          type: integer
        completed:
          type: boolean
      required: [count, completed]

    Favorite:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [surah, dhikr, verse]
        referenceId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, referenceId, title, createdAt]

    AddFavoriteRequest:
      type: object
      properties:
        type:
          type: string
          enum: [surah, dhikr, verse]
        referenceId:
          type: string
        title:
          type: string
      required: [type, referenceId, title]

    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        hijriDate:
          type: string
        type:
          type: string
          enum: [personal, islamic, reminder]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, date, type, createdAt]

    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        hijriDate:
          type: string
        type:
          type: string
          enum: [personal, islamic, reminder]
      required: [title, date, type]