# lib/api-zod/src/generated/types/favorite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Noor Islamic App API
 * OpenAPI spec version: 0.1.0
 */
import type { FavoriteType } from "./favoriteType";

export interface Favorite {
  id: number;
  userId: number;
  type: FavoriteType;
  referenceId: string;
  title: string;
  createdAt: Date;
}