# lib/api-zod/src/generated/types/updateAdhkarProgressRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Noor Islamic App API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAdhkarProgressRequest {
  count: number;
  completed: boolean;
}